Fix bugs in price parsing. Convert price from float32 to int

+++ types/types.go
@@ -24,5 +24,5 @@ type CraigslistItem struct {
 	ThumbnailUrl string    `json:"thumbnail_url",db:"thumbnail_url"`
 	IndexDate    time.Time `json:"index_date",db:"index_date"`
 	PublishDate  time.Time `json:"publish_date",db:"publish_date"`
-	Price        float32   `json:"price",db:"price"`
+	Price        int       `json:"price",db:"price"`
 }
